Changed the MessagesController to a ChatController, Added neccesary field in the Message entity, Added db migration, Fixed null fields in response issue.
