+++ 924-server/Entities/Chat.cs
@@ -3,7 +3,5 @@
     public class Chat
     {
         public int Id { get; set; }
-
-        public int NumberParticipants { get; set; } = 10;
     }
 }
