+++ 924-server/Entities/User.cs
@@ -5,4 +5,6 @@ public class User
     public int Id { get; set; }
 
     public string Name { get; set; } = string.Empty;
+
+    public string ProfilePhotoUrl { get; set; } = string.Empty;
 }
